# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TournamentGame = typeof TournamentGame[keyof typeof TournamentGame];


export const TournamentGame = {
  bgmi: 'bgmi',
  'free-fire': 'free-fire',
} as const;

export type TournamentMode = typeof TournamentMode[keyof typeof TournamentMode];


export const TournamentMode = {
  solo: 'solo',
  duo: 'duo',
  squad: 'squad',
} as const;

export type TournamentStatus = typeof TournamentStatus[keyof typeof TournamentStatus];


export const TournamentStatus = {
  open: 'open',
  live: 'live',
  completed: 'completed',
} as const;

export interface Tournament {
  id: number;
  title: string;
  game: TournamentGame;
  mode: TournamentMode;
  status: TournamentStatus;
  startsAt: string;
  /** @minimum 0 */
  prizePool: number;
  /** @minimum 0 */
  entryFee: number;
  /** @minimum 1 */
  slots: number;
  /** @minimum 0 */
  filledSlots: number;
  map?: string;
  organizer?: string;
  /** @nullable */
  roomId?: string | null;
  /** @nullable */
  roomPassword?: string | null;
  description?: string;
}

export type TournamentInputGame = typeof TournamentInputGame[keyof typeof TournamentInputGame];


export const TournamentInputGame = {
  bgmi: 'bgmi',
  'free-fire': 'free-fire',
} as const;

export type TournamentInputMode = typeof TournamentInputMode[keyof typeof TournamentInputMode];


export const TournamentInputMode = {
  solo: 'solo',
  duo: 'duo',
  squad: 'squad',
} as const;

export interface TournamentInput {
  /** @minLength 2 */
  title: string;
  game: TournamentInputGame;
  mode: TournamentInputMode;
  startsAt: string;
  /** @minimum 0 */
  prizePool: number;
  /** @minimum 0 */
  entryFee: number;
  /** @minimum 1 */
  slots: number;
  /** @minLength 2 */
  map: string;
  /** @minLength 2 */
  organizer: string;
  /** @minLength 2 */
  description: string;
}

export type TournamentUpdateStatus = typeof TournamentUpdateStatus[keyof typeof TournamentUpdateStatus];


export const TournamentUpdateStatus = {
  open: 'open',
  live: 'live',
  completed: 'completed',
} as const;

export interface TournamentUpdate {
  status?: TournamentUpdateStatus;
  /** @nullable */
  roomId?: string | null;
  /** @nullable */
  roomPassword?: string | null;
  /** @minimum 0 */
  filledSlots?: number;
}

export interface RegistrationInput {
  /** @minLength 2 */
  playerTag: string;
  /** @minLength 2 */
  playerName: string;
  /** @nullable */
  teamName?: string | null;
}

export type RegistrationStatus = typeof RegistrationStatus[keyof typeof RegistrationStatus];


export const RegistrationStatus = {
  confirmed: 'confirmed',
  cancelled: 'cancelled',
} as const;

export interface Registration {
  id: number;
  tournamentId: number;
  tournamentTitle: string;
  playerTag: string;
  playerName: string;
  /** @nullable */
  teamName: string | null;
  joinedAt: string;
  slotNumber: number;
  status: RegistrationStatus;
}

export type LeaderboardEntryGame = typeof LeaderboardEntryGame[keyof typeof LeaderboardEntryGame];


export const LeaderboardEntryGame = {
  bgmi: 'bgmi',
  'free-fire': 'free-fire',
} as const;

export type LeaderboardEntryTrend = typeof LeaderboardEntryTrend[keyof typeof LeaderboardEntryTrend];


export const LeaderboardEntryTrend = {
  up: 'up',
  down: 'down',
  same: 'same',
} as const;

export interface LeaderboardEntry {
  rank: number;
  playerTag: string;
  playerName: string;
  game: LeaderboardEntryGame;
  matches: number;
  wins: number;
  kills: number;
  points: number;
  trend: LeaderboardEntryTrend;
}

export interface DashboardSummary {
  playerTag: string;
  tournamentsJoined: number;
  liveMatches: number;
  totalPoints: number;
  winRate: number;
  nextMatch: Tournament | null;
}

export type ListTournamentsParams = {
game?: ListTournamentsGame;
status?: ListTournamentsStatus;
};

export type ListTournamentsGame = typeof ListTournamentsGame[keyof typeof ListTournamentsGame];


export const ListTournamentsGame = {
  all: 'all',
  bgmi: 'bgmi',
  'free-fire': 'free-fire',
} as const;

export type ListTournamentsStatus = typeof ListTournamentsStatus[keyof typeof ListTournamentsStatus];


export const ListTournamentsStatus = {
  all: 'all',
  open: 'open',
  live: 'live',
  completed: 'completed',
} as const;

export type ListRegistrationsParams = {
playerTag: string;
};

export type ListLeaderboardParams = {
game?: ListLeaderboardGame;
};

export type ListLeaderboardGame = typeof ListLeaderboardGame[keyof typeof ListLeaderboardGame];


export const ListLeaderboardGame = {
  all: 'all',
  bgmi: 'bgmi',
  'free-fire': 'free-fire',
} as const;

export type GetDashboardSummaryParams = {
playerTag: string;
};