# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Tournament } from './tournament';

export interface DashboardSummary {
  playerTag: string;
  tournamentsJoined: number;
  liveMatches: number;
  totalPoints: number;
  winRate: number;
  nextMatch: Tournament | null;
}