# lib/api-zod/src/generated/types/registration.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RegistrationStatus } from './registrationStatus';

export interface Registration {
  id: number;
  tournamentId: number;
  tournamentTitle: string;
  playerTag: string;
  playerName: string;
  /** @nullable */
  teamName: string | null;
  joinedAt: Date;
  slotNumber: number;
  status: RegistrationStatus;
}