# lib/api-zod/src/generated/types/tournament.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TournamentGame } from './tournamentGame';
import type { TournamentMode } from './tournamentMode';
import type { TournamentStatus } from './tournamentStatus';

export interface Tournament {
  id: number;
  title: string;
  game: TournamentGame;
  mode: TournamentMode;
  status: TournamentStatus;
  startsAt: Date;
  /** @minimum 0 */
  prizePool: number;
  /** @minimum 0 */
  entryFee: number;
  /** @minimum 1 */
  slots: number;
  /** @minimum 0 */
  filledSlots: number;
  map?: string;
  organizer?: string;
  /** @nullable */
  roomId?: string | null;
  /** @nullable */
  roomPassword?: string | null;
  description?: string;
}