# lib/api-zod/src/generated/types/tournamentInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TournamentInputGame } from './tournamentInputGame';
import type { TournamentInputMode } from './tournamentInputMode';

export interface TournamentInput {
  /** @minLength 2 */
  title: string;
  game: TournamentInputGame;
  mode: TournamentInputMode;
  startsAt: Date;
  /** @minimum 0 */
  prizePool: number;
  /** @minimum 0 */
  entryFee: number;
  /** @minimum 1 */
  slots: number;
  /** @minLength 2 */
  map: string;
  /** @minLength 2 */
  organizer: string;
  /** @minLength 2 */
  description: string;
}