# lib/api-zod/src/generated/types/tournamentUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { TournamentUpdateStatus } from './tournamentUpdateStatus';

export interface TournamentUpdate {
  status?: TournamentUpdateStatus;
  /** @nullable */
  roomId?: string | null;
  /** @nullable */
  roomPassword?: string | null;
  /** @minimum 0 */
  filledSlots?: number;
}